+++ poisson_approval/__init__.py
@@ -112,6 +112,7 @@ from poisson_approval.meta_analysis.binary_plots import BinaryAxesSubplotPoisson
 from poisson_approval.meta_analysis.binary_shortcuts import binary_plot_n_equilibria, \
     binary_plot_winners_at_equilibrium, binary_plot_winning_frequencies, binary_plot_convergence, \
     XyyToProfile
+from poisson_approval.meta_analysis.distribution_scores import plot_distribution_scores
 from poisson_approval.meta_analysis.ternary_plots import TernaryAxesSubplotPoisson, ternary_figure
 from poisson_approval.meta_analysis.ternary_shortcuts import ternary_plot_n_equilibria, \
     ternary_plot_winners_at_equilibrium, ternary_plot_winning_frequencies, ternary_plot_convergence, \